move table to right folder
