+++ src/plugins/serlo-table/renderer.tsx
